Rollback minor refactors
