+++ src/transactions/feed/detailContent/TransferSentContent.tsx
@@ -31,7 +31,6 @@ function TransferSentContent({ transfer }: { transfer: TokenTransfer }) {
   const recipient = getRecipientFromAddress(address, info, metadata.title, metadata.image)
 
   const { securityFee, dekFee, totalFee, feeCurrency } = usePaidFees(fees)
-
   const totalFromFeesInLocal = fees.reduce(
     (sum, fee) => sum.plus(fee.amount?.localAmount?.value ?? 0),
     new BigNumber(0)
