+++ src/transactions/feed/queryHelper.ts
@@ -13,7 +13,7 @@ import { getDynamicConfigParams } from 'src/statsig/index'
 import { StatsigDynamicConfigs } from 'src/statsig/types'
 import { vibrateSuccess } from 'src/styles/hapticFeedback'
 import { updateTransactions } from 'src/transactions/actions'
-import { transactionHashesByNetworkIdSelector } from 'src/transactions/selector'
+import { transactionHashesByNetworkIdSelector } from 'src/transactions/reducer'
 import { NetworkId, TokenTransaction, TransactionStatus } from 'src/transactions/types'
 import Logger from 'src/utils/Logger'
 import { gql } from 'src/utils/gql'
