+++ src/transactions/reducer.ts
@@ -1,8 +1,11 @@
+import { createSelector } from 'reselect'
 import { ActionTypes as ExchangeActionTypes } from 'src/exchange/actions'
 import { NumberToRecipient } from 'src/recipients/recipient'
 import { getRehydratePayload, REHYDRATE, RehydrateAction } from 'src/redux/persist-helper'
+import { RootState } from 'src/redux/reducers'
 import { Actions, ActionTypes } from 'src/transactions/actions'
 import {
+  ConfirmedStandbyTransaction,
   Fee,
   NetworkId,
   StandbyTransaction,
@@ -160,6 +163,68 @@ export const reducer = (
       return state
   }
 }
+
+export const standbyTransactionsSelector = (state: RootState) =>
+  state.transactions.standbyTransactions
+
+export const pendingStandbyTransactionsSelector = createSelector(
+  [standbyTransactionsSelector],
+  (transactions) => {
+    return transactions
+      .filter((transaction) => transaction.status === TransactionStatus.Pending)
+      .map((transaction) => ({
+        ...transaction,
+        transactionHash: transaction.transactionHash || '',
+        block: '',
+        fees: [],
+      }))
+  }
+)
+
+export const confirmedStandbyTransactionsSelector = createSelector(
+  [standbyTransactionsSelector],
+  (transactions) => {
+    return transactions.filter(
+      (transaction): transaction is ConfirmedStandbyTransaction =>
+        transaction.status === TransactionStatus.Complete ||
+        transaction.status === TransactionStatus.Failed
+    )
+  }
+)
+
+export const knownFeedTransactionsSelector = (state: RootState) =>
+  state.transactions.knownFeedTransactions
+
+export const recentTxRecipientsCacheSelector = (state: RootState) =>
+  state.transactions.recentTxRecipientsCache
+
+export const transactionsByNetworkIdSelector = (state: RootState) =>
+  state.transactions.transactionsByNetworkId
+
+export const transactionsSelector = createSelector(
+  [transactionsByNetworkIdSelector],
+  (transactions) => {
+    const transactionsForAllNetworks = Object.values(transactions).flat()
+    return transactionsForAllNetworks.sort((a, b) => b.timestamp - a.timestamp)
+  }
+)
+
+export const inviteTransactionsSelector = (state: RootState) =>
+  state.transactions.inviteTransactions
+
+export const transactionHashesByNetworkIdSelector = createSelector(
+  [transactionsByNetworkIdSelector],
+  (transactions) => {
+    const hashesByNetwork: {
+      [networkId in NetworkId]?: Set<string>
+    } = {}
+    Object.entries(transactions).forEach(([networkId, txs]) => {
+      hashesByNetwork[networkId as NetworkId] = new Set(txs.map((tx) => tx.transactionHash))
+    })
+
+    return hashesByNetwork
+  }
+)
 
 export const useGasFees = ({
   gasFee,
